requirements.txt
aiogram>=3.0.0
transformers>=4.30
torch>=2.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
python-dotenv>=0.19.0